Handle multiple taxes in invoice line
